[6CORE-007] checkstyle
